+++ packages/terra-clinical-item-collection/CHANGELOG.md
@@ -3,6 +3,8 @@ ChangeLog
 
 Unreleased
 ----------
+* Updated nightwatch test scripts
+* Removed nightwatch.config file
 * Converted component to use CSS modules
 
 1.0.0 - (June 28, 2017)